Fix ClassNotFoundException
The bug is that on devices without the flag enabled, while the tests
should be flagged (and so disabled), the test infra still scans through all the methods in the class annotated with @RunWith(BedsteadJUnit4::class). It hits ClasNotFoundException because some of our test util methods take sidecar classes in the function signature. Moving the functions that depends on the sidecar jar to a companion class.

Flag: android.app.appfunctions.flags.enable_app_function_manager
Test: Existing CTS
Bug: 376499142
Bug: 378809429
Bug: 378159453
Bug: 378609846
Bug: 378060534
Change-Id: I22a51fb679bdfdc41eca7b382480c9f6a4ae13e1