Run media session related tests with bedstead
This change fixes the failures reported
by below bugs by exempting certain test methods that
can't succeed when run with a user of type
`profile.MANAGED` because such users aren't a full
user.

Running the tests with bedstead and annotate them
to run for certain users makes sure these test are run
for these users when run with 'atest'. Before this change
the tests where run only for the owner.

Bug: 376412792
Bug: 378817823
Bug: 378520643
Bug: 378781441
Bug: 354813096
Flag: TEST_ONLY
Test: Part of CTS media tests
Change-Id: I650a6954423e972cf31023982f43c2572b0952fa