CtsLintChecker: report attribute location
When reporting an issue about the package name, the location of the
manifest tag was used. Be more precise by reporting the exact location
of the attribute value.

This is necessary as the presubmit linter relies on the exact line being
reported. If "package" was on a different line than "manifest" (which
is often the case), the warning might have been ignored.

Bug: 380139521
Test: atest --host CtsLintCheckerTest
Change-Id: I3da378b71457d1ba5a7db993656a57706db9e6c0