Relax some ActivityManagerProcessStateTests expects with waitFors
The WatchUidRunner is a bit racy and there are several points in the
ActivityManagerProcessStateTests that expect specific state that can
trip on the the timeliness of watch-uid events. When the tests were
expecting the cached state, there is sometime a Service event still in
the WatchUidRunner queue.

Also, updated WatchUidRunner#waitFor to allow for failing on specific
state to maintain test expectations (like the FGS state should not be
hit in the scenarios that disallow FGS starts).

Flag: EXEMPTED_testOnly
Bug: 370817323
Bug: 380237277
Test: atest ActivityManagerProcessStateTest

Merged-In: I7d55e39ea25a052ffa377daea9cc595f18466b16
Change-Id: I7d55e39ea25a052ffa377daea9cc595f18466b16