Fix parsing for APIs using generics
Generics parsing and codegen always assumed a single generic parameter.
We now have a Map in UserManager, which breaks this assumption.
Implement handling of multiple arguments.

This CL does not allow updating to the latest API yet. Codegen for
List<> will be fixed in a separate CL.

Bug: 381395375
Flag: TEST_ONLY
Test: m RemoteFrameworkClasses
Change-Id: Ic225d99ffa09bb398084e3f9a045b94b4fdfff4c