Replace deprecated AndroidTestCase for NetSecPolicy CTS tests
The AndroidTestCase class is deprecated and assumeTrue does not work
correctly. The NetworkSecurityPolicyTests report "FAILED" when there is
no internet connectivity instead of "ASSUMPTION_FAILED", resulting in
unnecessary flakiness in the test module.

This CL rewrites the CTS tests to use AndroidJUnit4 instead, where the
assumeTrue statement works as intended.

Flag: TEST_ONLY
Bug: 379911163
Test: atest CtsNetSecPolicyUsesCleartextTrafficUnspecifiedTestCases CtsNetSecPolicyUsesCleartextTrafficFalseTestCases CtsNetSecPolicyUsesCleartextTrafficTrueTestCases
Change-Id: Iefcb1e612e5a9e833c29430c2f2baf3818bb0ee6