DO NOT MERGE ANYWHERE: Loosen the requirement on tag
We need to override the session tag flag to true in test branch ap3a
which will break this test if device is running it with false value.
This will loosen the test so it only ensures that the metric is sent
but not check the session tag value difference caused by flag.

Bug: 381463488
Test: atest PerformanceHintManagerStatsTests CtsADPFHostTestCases
Change-Id: Ic3f0750ceb9770e6d8993ff9fa1c0b45323c6a99