Wait for lock screen focused before unlocking
The testRemoveCurrentProfileCanStartInputOnOtherUser case locks the
screen, removes the profile user, turns on the screen and unlocks. This
fails sometimes due to a race condition where the unlock is attempted
before the lock screen is focused. The fix waits for a non activity
window (lock screen) to be focused before trying to unlock.

Flag: EXEMPT bugfix
Bug: 377860468
Test: MultiUserMockImeTest#testRemoveCurrentProfileCanStartInputOnOtherUser
Change-Id: I208fdf7de8225cdbb684ad5229a2d5ce276fdc72