Ensure that activities launched in IntentTests are the ones from the intent.
Issue: When an activity is launched and is reset to the new task, a previous activity may also be resumed (e.g. in freeform mode), which updates the mLastActivity in the activity monitor. This makes it so that the activity monitor would report the previous activity and not the one launched.

Bug: 377583299
Change-Id: I8de5c02974c15794c2a7366b6e2c47d46a081916
Test: IntentTests
Flag: TEST_ONLY