ITS: Split the AWB check from AE check in test_multi_camera_switch
Most of the devices are skipping test_multi_camera_switch test in VIC mostly due to majority of the devices support ae_regions and not awb_regions.
Split the checks in this test and perform AE and AF check if
ae_regions is supported and AWB check only if awb_regions is supported.

Flag: TEST_ONLY
Test: run scene7
Bug: 379347590
Change-Id: I859aacbfb08a408524d75f9130a20792bf317f50