Add CTS tests that trigger a log list download
Test cases added:
- log list download triggered -> connects successfully to SCT-provided
  domain
- log list download triggered -> fails to connect to no SCT-provided
  domain (throws expected exception)

Pre test commands change the permissions of the parent directory of the
CT root directory, so as to allow the FileObserver to work properly.
They also delete any existing CT directory, as these tests are
specifically testing out the download flow.

On class tearDown, the CT log list directory is deleted so as to leave
the testing environment the same as when the test was initially run.

Bug: 378421935
Bug: 378424118
Bug: 378423774
Flag: TEST_ONLY
Test: atest NetSecConfigCertificateTransparencySctLogListTestCases &&
atest NetSecConfigCertificateTransparencyNoLogListTestCases

Change-Id: I676825c94d946e43d2a39b4f9511ffa1b2fe4aee