RESTRICT AUTOMERGE: Roll forward of "Make WindowUntrustedTouchTest multi-user and multi-display aware"
This reverts commit 78d94653f2a0976924deba945209674aa085dbf7.
Reason for revert: previous CL was mistakenly pointed as the culprit.

This changelist updates WindowUntrustedTouchTest to pass the display ID associated with the test activity (mActivity) to UntrustedTouchTestService. This allows UntrustedTouchTestService to create views with the correct window manager and window context.

Previously, this test failed when running for a secondary user on a secondary display (a visible background user on a secondary display). This failure occurred because DEFAULT_DISPLAY is not the main display for such users. Instead, their main display ID is assigned when the background visible user is created. This change ensures the test uses the correct display ID in all scenarios.

NOTE FOR REVIEWERS: this is a dirty where changes in WindowUntrustedTouchTestBase.java from the original CL had to be adapted.

Flag: TEST_ONLY
Test: atest CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest
Test: atest --user-type secondary_user CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest
Bug: 378456865
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:fe376e1608837793d7041e42c4466aa7275ba62f)
Merged-In: I16c381291939f9b38307faad0c3a58abc2760755
Change-Id: I3176972237691a7e259856e7b264d9bffbf31615