Add test for missing dependency role holder
Add a CTS test that tries to install an app when there is no dependency
installer service. Installation should fail with a "dependency installer
service not found" error.

Since we never unbind from the DependencyInstallerService, there could
be other tests where this service is bound and still present. To prevent
that from affecting this test, a new user is created for this test so
that we attempt to bind to a new service each time instead of reusing
any existing one.

Bug: 372862145
Test: atest PackageManagerShellCommandInstallTest
FLAG: android.content.pm.sdk_dependency_installer
Change-Id: I56ab50a5cf3a545fccae5cb66dad754ae96aa822