Fix SecurityStateManagerTest with clearCallingIdentity implementation
It doesn't seem like CTS can access the package info across user profiles (even with clear calling identity) unless the package is added under a queries tag in the manifest. So the default mainline and webview packages were added. Since the user can modify the config and add any packages they want, CTS may not always be able to get the package info. So the failing assert was changed to a warning log

Change-Id: I5e3b53fc65f26902d2909311c9fed5ad1bb46318
Flag: TEST_ONLY
Bug: 381582454
Test: https://android-build.corp.google.com/builds/abtd/run/L63800030008331154