Clean up IpcDataCacheTest calls
The IpcDataCache methods getDisabledState() and isDisabled() are
redundant.  In PropertyInvalidatedCache, getDisabledState() is just a
call to isDisabled().  This replaces all calls to getDisabledState()
with a call to isDisabled(), in preparation for removing the obsolete
test api.

Note that IpcDataCache already calls isDisabled(), so this is not a
new API for the tests.

Flag: EXEMPT refactor
Bug: 385198070
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
Change-Id: I2a68f4754c0eeee0ac695ec04c0783091dcb3393