Fix SplitPermissionAutoGrantTest for secondary_user_on_secondary_display
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

The fix is to take the displayId in the BySelector so that the
ALLOW/DENY button can be located on the secondary display.

Bug: 385358585
Flag: EXEMPT test fix only
Test: Tested on sdk_car_md emulator
Test: atest android.security.cts.SplitPermissionAutoGrantTest#testAutoGrant
Test: atest --user-type secondary_user_on_secondary_display android.security.cts.SplitPermissionAutoGrantTest#testAutoGrant
Change-Id: I26d9b7642519da7434a999212dbdd4a14354c9bd