Fix FileIntegrityManagerTest for secondary_user_on_secondary_display
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

Test was previously failing due to Caller not having REQUEST_INSTALL_PACKAGES. The fix is to grant the permission to the test running user explicitly instead of assuming the testing running user is always the current user.

Bug: 385357166
Flag: EXEMPT test only fix
Test: Tested on sdk_car_md emulator
Test: atest android.security.cts.FileIntegrityManagerTest
Test: atest --user-type secondary_user_on_secondary_display android.security.cts.FileIntegrityManagerTest
Change-Id: Iae016639a501501442f31847215f0595bbdc6388