Keep CPU running after CTS injects the sleep key
ADB connections are usually established over USB, which ensures that the
device does not suspend due to a kernel wake lock held by the USB
driver.

However, if ADB connections are established over TCP, there is no wake
lock, and the device suspends immediately after the screen turns off.

To fix this issue, this change manually acquires a wake lock before
the tests send the sleep key to the device.

Bug: 368153627
Flag: EXEMPT bugfix
Test: cts-tradefed run commandAndExit cts --include-filter "CtsWindowManagerDeviceActivity android.server.wm.activity.ActivityVisibilityTests"
Test: cts-tradefed run commandAndExit cts --include-filter "CtsWindowManagerDeviceDisplay android.server.wm.display.DisplayTests"
Test: cts-tradefed run commandAndExit cts --include-filter "CtsWindowManagerDeviceMultiDisplay android.server.wm.multidisplay.MultiDisplayKeyguardTests"
Test: cts-tradefed run commandAndExit cts --include-filter "CtsWindowManagerDeviceMultiDisplay android.server.wm.multidisplay.MultiDisplayActivityLaunchTests"
Change-Id: I566f32a58fcbdfa29b08928cba7457c9126b391b