Uncompress and align APK, DM, and SDM files.
The goal of SDM files is to avoid dex2oat on install. Therefore, the DEX
files in the APK need to be uncompressed and aligned. Otherwise, dex2oat
has to be run, to extract DEX files.

The ODEX file in the SDM file must be uncompressed and aligned so that
it can be loaded by Bionic's dlopen.

The VDEX file in the DM file and the ART file in the SDM file should be
uncompressed and aligned so that ART can mmap them directly, to avoid
the runtime cost. Those files being compressed may be supported but is
not expected.

Bug: 377474232
Test: atest CtsCompilationTestCases
Change-Id: If4a0bd471df13043bd7ad883fbb63027c239baf5