WindowInputTests: log insets being used move-and-tap test
In b/380004295, it appeared that the insets were being ignored by the
test, causing flakes. The flakes seem to have passed now, and I can't
reproduce the problem locally, so just add a log for the insets being
used (and the types used to retrieve them), to give us more insight if
it happens again.

Bug: 380004295
Test: $ atest 'CtsWindowManagerDeviceInput:android.server.wm.input.WindowInputTests#testMoveWindowAndTap'
      and check the log message appears in logcat
Flag: TEST_ONLY
Change-Id: Ib259c9eb0c9ac43ab4743b826f6dd1479c592545