Fix test finding incorrect scrollables
On devices with scrollables present outside of the test activity, the
test may not be able to find the view it is looking for. The issue is
fixed by only looking through scrollables in the activity area.

Bug: 387388108
Test: atest
android.preference.cts.PreferenceActivityFlowPortraitTest#listDialogPortraitTest
on cf_gwear_x86-trunk_staging-userdebug and
gcar_x86_64-trunk_staging-userdebug
Flag: EXEMPT test_only

Change-Id: Idad64d4b004f1a91065562d57cc48cf24f22d3bd