Check remote_views_proto flag for proto tests
Adds check for proto flag if isProtoTest is true. If the proto flag is
not enabled, skips the test. This is similar behavior to
CheckFlagsRule but that rule does not support conditioning the flag
check on a parameter value.

Fixes proto tests that are failing on targets without the flag enabled.

Bug: 386446568
Bug: 387430907
Bug: 386445553
Test: atest CtsWidgetTestCases:android.widget.cts.RemoteViews{,Activity,SizeMap,FixedCollectionAdapter,Recycling,ThemeColors}Test
Test: on cf_gwear_x86-next-eng
Flag: android.appwidget.flags.remote_views_proto
Change-Id: I0922cd3bf040107160e131a20a99d8bb070f2050