Fast fail test is no valid SIM is present.
If there is no valid sim in the device, mPhoneAccountHandle will have an
id of "E" which is going to cause mTelephonyManager to be null.  This will
show up erroneously as an NP when "getHalVersion" is called.

Ensure we fail out of the test if no sim is present.

Flag: NONE test only fix.
Fixes: 387244577
Test: This is a test fix.
Change-Id: Iff571d1118bbd6eda9fc09d2b7d98b31940106b9