Create test to ensure requested exposure values are respected.
On some devices, the exposure and gain metadata reported for captures
with auto exposure enabled do not match for a manual session with those
same values (the image will appear darker). Images captured with the
same exposure and gain settings as an auto exposure session should look
rougly the same.

This test takes two captures - one with AE on and another with AE off.
It then splits the image into 16x16 block regions and computes the SAD
between the two images for each block. The block with the maximum SAD
is compared against a threshold, and if it exceeds this threshold the
test fails. To reduce the impact of noise, the two captures are gaussian
blurred.

Bug: 301570451
Bug: 385220060
Test: Ran test on two devices, one with the issue and one without.
Flag: EXEMPT ITS test
Change-Id: Iafb258ad2df516cb6b916dcbc25575948718bb20