Add AvailableIntentsTest for ACTION_MOTION_PHOTO_CAPTURE.
On devices where there is an implicit intent handler for this action
(which is not required to be supported), the intent must be handled
only by the pre-installed camera app.

On devices without a pre-installed camera app handling the action, newly
installed apps should not hijack the implicit intent.

Bug: 384790695
Test: Ran test on physical device.
Flag: com.android.providers.media.flags.motion_photo_intent
Change-Id: Ia24372290e3bf9c67f778309aec3d66ecfc518df