Suppress removeKeyPair error
This error occurs within a `finally` block, so it possibly overshadows real issue causing trouble in this bug (in the main `try` block of the test). This commit wraps this misleading error and only logs it instead of throwing, giving chance for the actual problem to stop lurking in the shadows and reveal itself.

Local reproduction is impossible, launching a class with all its tests leads to all of them passing all the time. Meaning there is a part of the state not cleaned up properly somewhere before, in other classes of the module.

Bug: b/333230523
Flag: TEST_ONLY
Test: atest CtsDevicePolicyTestCases:android.devicepolicy.cts.CredentialManagementAppTest
Change-Id: I11953ad2347ed24ed526583e93c115d3964437d7