Fix testRequestFrontendNoFrontendAvailableAndResourceOwnerRetention
The original test seems to avoid using real tuner. However, this cause
the release frontend process stop in the middle because the
mFronendHandle is null. The test will always fail. This CL change the
test to use the real tuner instead and following normal apply frontend
process.

Bug: 389858883
Test: atest TunerTest
Flag: TEST_ONLY
Change-Id: I2e79ef281f33eebcef3aeb610d667c2e621936b4