ITS: Collect cameraservice watch dump output.
This change enables starting watch cameraservice on default camera and
JCA packages in order to monitor the tags. This output is helpful in
determining the various control parameters which differ during
caputures.
The watch cameraservice dump output is saved in a text file which is
available in the test output directory.

Bug: 380892830

Test: Tested on pixel device

FLAG: TEST_ONLY
Change-Id: I9f267430771745287ff00edbd27aa165a61d63d9