ITS: Add thresholds for wb and brightness checks.
Check if the mean white balance and brightness diff between JCA and
default captures falls within the determined threshold level or not.
If the device does not meet the criteria, then log the error message and
raise error with that error message. It is concated with
NOT_YET_MANDATED message so even if the test fails, it will be marked as
Fail* on devices.

Bug: 380892830

Test: Tested on pixel device
FLAG: TEST_ONLY
Change-Id: Ifd21a1acf9ed0d3a4729234148fba4a20ec1d131