ITS:Update setup to read rotator controller params
Currently we only read the rotator controller and id if the test bed is
sensor_fusion. With newly added gen2 rig testbed which needs both
rotator and light controller, the setup skips the part to read the
params. Read these params if present so that the gen2 rig hardware can
be controlled with minimum changes.

Bug: 380892830
Test: Tested with gen2 rig.
FLAG: TEST_ONLY
Change-Id: Ia2119d3053a9c64572637e304f4b160f1f20f7dd