Fix test flakiness in testOverlappingActivity.
The test testOverlappingActivitySandwichDuringAnimation_DoesNotBlockTouc
hes was flaky, with a failure rate of 3.59%, because it was not waiting
for the app transition to finish before sending touch events, causing
the touch events to be dropped. This change adds a wait for app
transition idle before sending the touch events, which ensures that the
activity is ready to receive touch events.

Fix: 389826364
Test: atest CtsWindowManagerDeviceActivity:ActivityRecordInputSinkTests#testOverlappingActivitySandwichDuringAnimation_DoesNotBlockTouches
Flag: TEST_ONLY
Change-Id: I3771b945135b81f9b8b484ebf50480b417e1ca00