Fix AccessibilityTextActionTest failing on some devices.
The test failed on devices that did not start app windows at
(0, 0) because it was comparing window coordinates to
screen coordinates in order to calculate a scroll offset.
Updates the test to instead always use the same coordinate
system to calculate a screen offset.

Failing behavior was reproduced on cuttlefish emulator
and is fixed by this change locally.

Test: atest android.accessibilityservice.cts.AccessibilityTextActionTest#testTextLocations_textOutsideOfViewBounds_locationsInWindowShouldBeNull
Flag: android.view.accessibility.a11y_character_in_window_api
Bug: b/389007334
Change-Id: Id7c647a7a0824940091ac30ed9cfc6992bbe87b0