Remove aaudio testBasicMMapOutput from known failure.
The failure was originally added to known failure as there was not a way
to report if MMAP is supported or not on a particular device. This has
been resolved by the new API AAudio_getPlatformMMapPolicy. The test
currently uses the new API to decide if MMAP should be used or not. It
should pass when there is no issue with the DUT.

Bug: 279765104
Test: make
Flag: EXEMPT bugfix
Change-Id: I3ffcb7b630e42489f87b39b131f68be8bdb736f3