Optimize and extract hasOverlay in CTS navbar util
In [1] the CTS navbar util GestureNavSwitchHelper was improved to also
switch between three button nav and gesture nav, as well as properly
check the current navigation mode. The latter was slightly inefficient
as it relied on parsing the full dump output of the
OverlayManagerService.

This optimizes the flow to only check the state of the selected
navigation mode overlay, as well as extracting it. This allows some
early assertions to determine whether the device does have the two
navigation mode overlays installed.

  [1]: I87e87cd30e45f3eb73b009a13cffb1676e067590

Flag: EXEMPT test improvement
Bug: 377912666
Test: atest InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_gestureNav
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_threeButtonNav
Change-Id: Idc5a8b6990de77ca73b731934398afcd04ae55ad