Inline isGestureMoveWithOverlay
In [1] isGestureModeWithOverlay was introduced as an alternative to
isGestureMode, to fix problems with verifying this state on instant mode
test runs, and on some particular devices. The former variant was left
for existing tests, to avoid potential breakages.

Now that the new variant has been verified to work, it can be inlined as
the single remaining one.

  [1]: I87e87cd30e45f3eb73b009a13cffb1676e067590

Flag: EXEMPT cleanup
Bug: 377912666
Test: atest InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_gestureNav
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_threeButtonNav
  FullScreenMagnificationGestureHandlerTest#testTapNavigationBar_zooming_keepZooming
  FullScreenMagnificationGestureHandlerTest#testSwipeUpFromNavigationBar_zooming_zoomOut
  StylusHandwritingTest#testStylusSession_stylusWouldNotTriggerNavbarGestures
  StylusHandwritingTest#testStylusSession_fingerTriggersNavbarGestures
Change-Id: Iea07b138a625c9f33bf5960def613bc2dd64576d