Adds CTS Tests to verify DeviceState properties are being used
Tests that device state identifiers that are present in various
overlay configuration values (such as foldedDeviceStates or
deviceStatesOnWhichToSleep) have the correct matching properties on
the DeviceState objects which match the identifiers.

This will verify that devices are being configured to match
and work with the updated DeviceStateManager API design.

Flag: TEST_ONLY
Test: DeviceStateManagerTests
Bug: 392129966
Change-Id: Id08023660f029208357b5a4467e08ae9f1c8a40d