DO NOT MERGE Title: Fixing the CtsDevicePolicyTestCases clone user related test cases
Description: 4 failure cases need to be marked as assumption failure based on the feature availability android.software.device_admin
             The line - @EnsureDoesNotHaveUserRestriction(CommonUserRestrictions.DISALLOW_MODIFY_ACCOUNTS) for the method is not required
because the test method itself already handles the restriction explicitly within its logic as The test method already
applies the DISALLOW_MODIFY_ACCOUNTS restriction and also clears the restriction at the end of the test.
Since the method itself takes care of both applying and clearing the restriction, the line EnsureDoesNotHaveUserRestriction becomes redudant.
The test does not need to check whether the restriction is already in place at the start, because the test is
explicitly setting the restriction itself as part of its flow.

Based on the test method- removeAccount_withAccountManagementDisabled_throwsException, the primary goal of the test seems
to be ensuring that once the Account Management policy is applied , account removal is prevented (via the expected exception).
The test is also focused on checking whether the policy is enforced correctly - specifically whether account removal is prevented when
the AccountManagement policy is disabled. This falls under Policy enforcement and not just setting the policy.
Since the focus is on verifying the behavior of the system after applying the policy (i.e., enforcing the restriction),
the more appropriate annotation would be - PolicyAppliesTest rather than CanSetPolicy.

Test: run cts -m CtsDevicePolicyTestCases --abi arm64-v8a
Bug: 391920539
Change-Id: Id88d77f3da72852fbf71775af70201e87744fecd