Use "additionalUser" to make test run on HSUM devices
Secondary user on Automotive HSUM devices is the first human user as
well as other users. We want the test to install an app only on the
other user, exluding the first human user. This can be achieved by using
"additionalUser". On HSUM devices, this will be a different human user,
and on non-HSUM devices, it will be the guest user, different than the
primary user (user0)

Test: atest CtsPackageUninstallTestCases
Bug: 391038480
Flag: EXEMPT: Test fix only
Change-Id: If5653218c25495b0e765aee44bf5444132b918e6