Replace the CDM association flow in the VDM setup
with the new CDM TestApi to skip role grants.

Keeping the old implementation for SDK level <= 35 because the
tests in PermissionModule need it.

Disabling the CDM tests for the VDM profiles because the CTS package
doesn't qualify for the roles anymore and cannot create the
associations

Fix: 391599609
Test: atest
Flag: EXEMPT test improvement

Change-Id: Iaaf0534401b6163836961ddd6c218557ca2ceefc