Wait for IME crash when removing its user profile
Previously in testRemoveCurrentProfileCanStartInputOnOtherUser the
profile user's IME session was closed before the profile was removed,
and then it was checked that onDestroy was registered for that IME.

As closing the IME session also closes its event stream, this sometimes
leads to test failures when the onDestroy event (caused by the user being
removed) is received after the stream was closed, failing the assertion.

The reason for closing the IME session before removing the user is that
as part of the close flow the IME settings will be deleted from its
ContentProvider. However, if the user profile of that IME is removed
before this happens, the delete call throws an exception as it cannot
find the ContentProvider authority of an IME of a deleted user.

The onDestroy assertion was replaced by awaiting an IME process crash,
with the reason that the user was stopped. The close flow exception is
handled by introducing an setSuppressDeleteSettings option on the
ImeSettings, to avoid calling the delete command in such a case. As the
user is fully removed, any data tied to it should already be cleaned up.

Flag: EXEMPT testfix
Bug: 385227171
Test: atest MultiUserMockImeTest#testRemoveCurrentProfileCanStartInputOnOtherUser
Change-Id: Ia5b74529be6b532a2235369d4efb7c85befbdcee