Skip BAL tests on HSUM devices
BAL is relevant for devices with user interaction. The test cases assume
that the user can switch to a home app, which is not present in HSUM
mode.
It also catches cases where the Intent cannot be resolved. This might
need some additional investigation, because I was not able to reproduce
the errors.

Flag: com.android.window.flags.bal_strict_mode_ro
Bug: 393058532
Test: atest StrictModeTest
Change-Id: I57fb1603095579c76c897590eff5afe09cbfee12