Add EvemuDevice class to replay EVEMU recordings in tests
The EvemuDevice makes use of our exising VirtualInputDevice infra.

To do this integration, we need to separate the EVEMU files into two
commands for device registration and for event injection. We parse the
EVEMU file for the first event 'E' line, which we use to demarcate the
two sections.

Bug: 385141667
Bug: 389901828
Test: Presubmit
Flag: TEST_ONLY
Change-Id: Ib00c7a14422c276901f161cd817bb28fc8909408