Annotate MTE tests correctly
A few of the memory tagging extension tests for Advanced Protection were
not annotated correctly to show a dependency on the MTE policy flag
in the Android Enterprise codebase.

This was not caught earlier because Pixel devices did not declare
support for MTE previously.

Fix this by annotating the tests.

Bug: 394162745
Test: atest CtsSecurityTestCases:android.security.cts.advancedprotection.MemoryTaggingExtensionTest
Flag: android.app.admin.flags.set_mte_policy_coexistence
Change-Id: I777cd1464ddf61e7ce4dff9f814bcc374c515fcd