Don't wait for window geometry to stabilize when waiting for input
When waiting for an activity to be ready for input injection, we
first ensure that all app transitions on the display are complete
to account for animations. We then wait for the window to be on top
and not obscured. There is no need to wait more time for the geometry
to stabilize here because the transitions should already be complete.

Bug: 385141667
Change-Id: I1468ab1dde971ea3aef913ef40836f9988e5aa73
Test: Presubmit
Flag: TEST_ONLY