Get disallowed system apps with correct user
The API getDisallowedSystemApps is supposed to be called with the user
to be provisioned, not its parent. Fix this in two tests.

I verified this looks correct by checking production usage of the API in
managed provisioning. It calls it with the user id of the profile, and
not the parent as well.

Bug: None
Test: btest android.devicepolicy.cts.ProvisioningTest#createAndProvisionManagedProfile_removesNonRequiredAppsFromProfile android.devicepolicy.cts.CreateManagedProfileTest#createManagedProfile_removesNonRequiredAppsFromProfile
Flag: EXEMPT TEST_ONLY
Change-Id: I80af1a6c46fc96ecaf9ad743d8a334249169a4ec