Fix permissions on 2G AAPM CTS test
The refactor in ag/30980540 caused this to stop working. The calls to
ShellIdentityUtils reset the permissions, which means the test would no
longer have the MANAGE_ADVANCED_PROTECTION_MODE permission when
setAdvancedProtectionEnabled() was called.

Instead of using ShellIdentityUtils, we just acquire all the needed
permissions in the setUp() method.

Bug: 394483945
Test: atest CtsSecurityTestCases:android.security.cts.advancedprotection.DisallowCellular2GTest -- --abi x86_64
Flag: TEST_ONLY
Change-Id: I447c28f79098679bb40dc6a7190953fef9ba228c