[PM] Add scroll behavior for CtsSuspendAppsTestCases
The button may not be shown in the smaller device. Add the scroll
behavior in the test case.

MUST_SLEEP: It is used to dump Window Hierarchy when the test fails. It should be fine.

Flag: TEST_ONLY
Test: atest CtsSuspendAppsTestCases
Bug: 396298952
Merged-In: I828fe1de399a02f624c4a5f254b417e4c0f0bdaa
Change-Id: I828fe1de399a02f624c4a5f254b417e4c0f0bdaa