Resolve flaky call audio focus CUJ tests
Ensures that audio routing ops are completed at the end of each test. In
the failures for the outgoing tests, we see that there's overlap in call
audio routing state. For incoming test failures, we find that devices
that have vibration enabled will not try to acquire audio focus.
Instead, this will only happen once we explicitly move to active focus.

Bug: 397514952
Test: atest CtsTelecomCujTestCases:CallAudioRouteTest
Flag: EXEMPT test only
Change-Id: I0a2a36d669258094bc8b33c2df8fdbd96e5f3402