Replace flag check with sdk check
Now that Bakalva has been finalized we can do a SDK check as we know for
sure that the API is in Bakalva. This allows the lint tool to understand
that hte API call is safe. (Long term the lint tool should be taught
about flags, but we're not there yet)

Flag: EXEMPT Baklava Finalization
Bug: 395811473
Test: m out/soong/.intermediates/cts/tests/camera/cts-camera-performance-tests/android_common/lint/lint-baseline.xml
Change-Id: Iccf654a8ed057529334d4b17999f8a57e3d779c5