Remove delegation scopes after test
A speculative attempt to fix some flakiness in
testDelegatedCertInstallerDirectly: this test is only flaky on some
branches, and failure behaviour indicates the test app has device ID
access even though it shouldn't. The hypothesis is that another test
case granted it CERT_INSTALL delegation. Delegation scopes are stored
in the DevicePolicyData, instead of the more typical ActiveAdmin
data structure, so its clean up process could be different. However
since userId should not be reused within a boot session, it's not
obvious why a test app getting CERT_INSTALL in one work profile will
end up leaking that privilege to another test run which will be
in a different user.

There is one instance of delegation granting not followed by proper
cleaning up. So fix that in case this helps.

Bug: 390078062
Test: MixedProfileOwnerTest#testDelegatedCertInstallerDirectly
Flag: EXEMPT test fix
Change-Id: I388f2e1eb7bcb2e9c02f531b457fe0faa6e3d9a1