Filter returned entries and add "binder_" package prerix
Reporting all the Binder services that exist in the device resulted in
excessive growth of the data to handle.

Reduce the data size by introducing a Set that keeps the list if
services being monitored. Every service in the Set is reported
regardless of its availability. For this reason, the new `is_available`
field is also introduced and `interface_descriptor` is reported as an
empty string if the service does not exist.

Test: atest CtsEdiHostTestCases:android.edi.cts.ServiceDeviceInfo#testCollectDeviceInfo
Bug: 397599218
Bug: 383274773
Change-Id: I25d44bc5fa3171bd1266f280fe0f973c82f39ea3
Signed-off-by: Alessio Balsini <balsini@google.com>