Set the view back to non-selectable
**Root cause**
When the TextView has a range of selection, by default it will show a
popup menu in the onPostExecute of the AsyncTask. However, when the time
it tried to add the PopupMenu, the associated activity has been
destroyed.

**Solution**
Set the view back to non-selectable in attempt to wait and close the
PopupMenu.

Bug: 396983947
Test: atest AccessibilityEndToEndTest
Flag: EXEMPT test only
Change-Id: I299dba25c189af7bc4c55ec4f6d90a843201e3d0