Remove explicit modification of privacy indicators flag
Android 16 introduces a new restriction that only allowlisted
DeviceConfig flags can be modified by the shell user. The
camera_mic_icons_enabled flag is not on the allowlist since it
was only intended to be used during the initial feature development
and should be enabled by default. This commit updates all tests
that currently modify this flag to only verify that the flag
is enabled.

Bug: 390242290
Flag: android.security.protect_device_config_flags
Test: atest RecognitionServiceMicIndicatorTest
Test: atest HotwordDetectionServiceBasicTest
Change-Id: I25208ab1f65d130bfd6caa1eafc22c11ec99ec47