Use LockScreenSession in MultiUser IME tests
Previously we were using Bedstead to manage the lockscreen state in
MultiUser IME tests. However, the sleep method only waits for the screen
to be turned off, which might be too early. This would lead to removing
the user before Keyguard is fully shown.

Instead this uses LockScreenSession which will reliably wait for the
keyguard state to be updated before proceeding.

Flag: EXEMPT testfix
Bug: 385227171
Test: atest MultiUserMockImeTest#testRemoveCurrentProfileCanStartInputOnOtherUser
Change-Id: Id2308c4ca208b4f82ba9686adfdabb572a634e2e