test_exposure_keys_consistent: Take average SAD of entire image.
Do not look at the SAD of individual blocks within the image. Instead,
take the SAD of the image and divide it by the number of pixels. This
relaxes the test to allow for regional differences in the image, while
still enforcing that the two images are mostly similar in brightness.

Flag: EXEMPT not applicable
Bug: 391499247
Test: Ran on failing physical device
Change-Id: I2e716187f41b89d924bc1791d3e3dbf02ef14eed