Prevent WindowCtsActivity to restart on rotation
For devices that have fixed orientation by setting either feature
android.hardware.screen.landscape or android.hardware.screen.portrait,
testSetFitsContentForInsets_displayCutoutInsets_areApplied seems to be
failing due to that the Activity is being destroyed on an unhandled
configuration change. The forced rotation of the test alters the
smallestScreenWidthDp to half the screen widths dp.

This patch expands the configChanges flags in the manifest to also
include smallestScreenSize in order to prevent the activity from being
restarted during testing.

Test: atest CtsWindowManagerDeviceTestCases:android.server.wm.WindowTest#testSetFitsContentForInsets_displayCutoutInsets_areApplied
Bug: 396164953

Change-Id: I85b110725ae3aa7e04cb82701d7e815ffd2828c7