Update CarrierPrivilegesUtils exception message to clarify potential cause
When Phone process is halt or rebooting, CarrierPrivilegesUtils is not able to update carrier privileges for the CTS test cases. The original exception message misleads us to think this is a bug in CarrierPrivilegesUtils itself.

This change updates the exception message to make it clearer what the potential causes of the failure are. It helps to lead us to the right direction to locate the real root cause.

Bug: 391721593
Test: atest CtsCarrierApiTestCases
Flag: EXEMPT (minor debug message update)
Change-Id: I015e042f7556f8e8b99e33301173cb607b21b317