DSDA: Drop HELD with ACTIVE + RINGING nonholdable
When we have a held and active nonholdable call and we receive an
incoming (nonholdable) call, ensure that the held call is properly
disconnected. Telephony wasn't properly ensuring that this happened and
we ended up in a scenario with two held calls and an active incoming
call. Moving this logic up into Telecom to handle this explicitly as
part of sequencing will resolve this.

Bug: 401033055
Test: atest
CtsTelecomCujTestCases: CallSequencingManagedHoldRestrictionTest
Flag: com.android.server.telecom.flags.enable_call_sequencing
Change-Id: I96c8c192658c8918d14a8411f954a22cab4fa2ec