Modify BackupPreparer to support non-system users in HSUM mode
In Headless System User Mode (HSUM), the user interacting with
the device (the primary user, often user 10 or higher) is distinct
from the system user (user 0).

BackupPreparer previously hardcoded interactions with `bmgr` to target
the system user (user 0). This caused issues when running tests in
HSUM mode, as the tests might be executing under the context of the primary
user, but the preparer was attempting to check/modify settings for user 0.

This change updates BackupPreparer to be aware of the current user.

Bug: 377460063
Test: atest ShortcutManagerBackupTest on phone and desktop
Test: atest CtsBackupHostTestCases on phone and desktop
Flag: TEST_ONLY
Change-Id: I2b2e480693e57b0f84619cd23cb3446951bfa1ee