ITS: create workaround for JCA tests to ensure image is saved before exit
Details: certain captures may take longer to be saved successfully. The JCA image capture success snackbar may appear and disappear before the image is saved, so it's not an accurate representation of when the activity can be exited safely. Added a TODO and an arbitrary timeout as a workaround until a permanent fix can be added.

Flag: TEST_ONLY
Bug: 404350495
Test: tested by partners and ran test_auto_flash locally
Change-Id: Ifdc7dcb00fe4e557b3633b5c9363c60bf6f3652a