Fix a race condition in two MediaRouter2Test cases
Both tests were assuming they could safely register an event listener
after taking an action that would trigger the event, and not have it
fire. Then later they took a different action, and would assert that
it did not cause the event to fire. But sometimes we got unlucky and
the event (caused by the first action) would end up firing, and the
tests would fail.

Test: atest MediaRouter2Test
Flag: TEST_ONLY
Fixes: 403913552
Fixes: 399019050
Bug: 291800179
Change-Id: I0bbaebffc4a650eb5502b284029c5bfe188fe309