Improve reliability of the CallDiagnosticService audio route test.
This test used to latch on the call audio state reported to the CDS and
compare it immediately.  This has proven to be flaky.

Instead, the CtsCallDiagnosticService will now track all of the audio
route changes made while it was bound.  We already wait for the ICS to
report that the audio route is speaker, which ensures that the session
which performs this update has completed.  We disconnect the call and wait
for that state change.  At this point we can be assured the CDS has
received any audio route changes it should have, so we check to be sure
one of them was a transition to speaker.

Flag: NONE test fix.
Fixes: 401559693
Test: Ran test multiple times to confirm success.
Change-Id: Ie523bbc7a84827153e15e5313cb516f7025ef9e0