Use assume to bypass test for devices w/o telecom
Originally every test case in the class extending BaseAppVerifier need
to add if condition to check the value of mShouldTestTelecom to bypass
the tests for devices w/o telecom. This results in a lot of duplicate
code and it is very easy to miss adding the if condition when adding new
tests. Use the assumeTrue to bypass the a better way to bypass.

Bug: 405883695
Test: run test on the TV dev device
Flag: EXEMPT test fix
Change-Id: Iefb4cf3830d5e22c64ac1c8527de149fd2b3a6df