Minor update on how the "clear all" button is searched.
On certain foldable devices, the entire notification pane may be
considered "scrollable", hence leading to dismissing of the pane
and the "clear all" button never being found. Update this logic to
only fetch the scrollable object in the heirarchy if the "clear all"
button is not initially found.

If there are many notifications, the button may not be visible so this
is simply a best effort. A follow-up change will be introduced to
better handle the clearing of notifications, especially for foldable
devices.

Bug: 403150807
Test: atest CtsSyncAccountAccessOtherCertTests
Flag: EXEMPT bugfix
Change-Id: I09226e2d2207169fb58a160dcd851b2f5260db84