SELinuxNeverallowRule: better ensure uniqueness of stable ID
CtsSecurityHostTestCases test runs are failing with the following
error message:

  RUNNER ERROR: Module arm64-v8a CtsSecurityHostTestCases only ran 1759 out of 1760 expected tests.

This is because the function getStableId() can produce the same
string from different SELinuxNeverallowRules, for example:

  neverallow { domain -init -vendor_init } { graphics_config_prop }:property_service set;

and

  neverallow { domain -init -vendor_init } graphics_config_prop:property_service set;

both will produce the same result from getStableId(), but produce
different SELinuxNeverallowRules.

Fix this issue by ensuring that the length of the test name exactly
matches the length of the SELinux rule. This will prevent string
cleanup from inadvertantly duplicating an existing test name.

To help prevent regressions, we introduce a new assert which
detects SELinuxNeverAllowRules with duplicate unique names.

To keep the length of the test name as short as possible, make
an additional change which reduces unnecessary whitespace.

Change-Id: If0c4a3bd0b8fb8365f220116023bc3fe9a2b3dfe
Test: atest CtsSecurityHostTestCases
Flag: TEST_ONLY
Bug: 407990681