(Reland) Add selectOneRoute_closeAndOpenDialog_routeStillSelected test
Reland context: the original CL ag/32641740 is not compatible with the
new output switcher redesign (b/408304744), and the presubmit and
postsubmit bots use different flag values. This CL is the same as the
original, but adds a call to assume the flag is off to avoid failures
as in b/407841928.

In addition to this new test case, this CL contains some new helper
methods and also support in the StubMediaRoute2ProviderService
callback proxying system to allow test cases to verify that the right
events are fired, but still take advantage of its default
implementation which does things like calling notifySession{Created,
Updated, Released}.

Test: atest OutputSwitcherTest
Flag: TEST_ONLY
Bug: 399188388
Change-Id: I87f7af5acfdbcabb2b03a84e8d9e5477876fe507