TextureViewTest: initial mSurfaceUpdatedCount could be non-zero
CaptionBar is drawn in desktop mode, and that's delegated to WM shell
done by sysui. Before that's being picked up, newly added TextureView
initial layouting could spot a dirty DecorView, ending up with an extra
onSurfaceTextureUpdated, advancing the mSurfaceUpdatedCount to 1.

Since the test is being strict about not dropping any of the first
frames drawn, this change just relaxes to allow the initial count being
non-zero. As long as the 2 affected instances, testFirstFrames and
testGetBitmap, have waited for onEnterAnimationComplete callback, the
initial surface update count is stable and can be used as a baseline to
further perform the tests.

Bug: 407035459
Test: atest CtsViewTestCases:android.view.cts.TextureViewTest
Flag: EXEMPT bugfix
Change-Id: I98b729c213ee47a79432fa994f7df401784e422c