videocodec: update testing bitrates for fullhd encoding
In vbr mode, format shaper to meet minimum quality requirements is
updating the configured bitrates to a new value. All the bitrates that
are used in roi test is being updated to a common new value. The bdrate
computation requires at least 4 distinct bitrate-quality points. As the
bitrates are updated to a common value, the 4 distinct points that were
chosen by the test are updated to a common point and thus failing the
test.

This change adds more entries in to the bitrates list. Also add CBR mode
to the validation list.

Bug: 400812394
Test: atest CtsVideoCodecTestCases:VideoEncoderRoiTest
Flag: TEST_ONLY

Change-Id: Ia945401476e58a22bbd6dd5f8b0bbaaec8316154