Improve the BarringFactor Check
It has been reported that some networks are using
a barring factor of 0 (which means that all UEs are
barred), for a certain amount of time. While it was
expected that this would actually result in unconditional
barring, networks are allowed to do this. Accordingly,
this is an opportunity to improve the check to support the
standard-defined values while also improving the check to
not allow values that are not among the 16 enumerated
percentages that are allowed for uac-barring or ac-barring.

Bug: 408976964
Test: atest CtsTelephonyTestCases
Flag: EXEMPT bugfix
Change-Id: Id15b06834a9a1b06e9fca5233d0e8e06a817066c