mediav2: update resource consumption threshold needs at perf points
The resource consumption at all performance points is expected to be
at least 60%. This requirement may not be valid especially for
performance points covering resolutions like 480p, 576p, 720p, ... .
The test is now updated to check for a performance occupancy threshold
that is scaled by a factor of
currPixelsProcessedPerSec / peakPixelsProcessedPerSec.

Here currPixelsProcessedPerSec corresponds to pixels processed per
sec for the performance point under test and peakPixelsProcessedPerSec
corresponds to pixels processed per sec for the first performance point.

Bug: 379220929
Test: atest CtsMediaV2TestCases:VideoDecoderAvailabilityTest
Test: atest CtsMediaV2TestCases:VideoEncoderAvailabilityTest
Flag: TEST_ONLY

Change-Id: I8697fa61e1d214fa67d50d2926a3516d80ea4ffb