AudioFocusTest: ensure volume setting to prevent flakes
When testing AudioManager.shouldNotificationPlay method,
actively set the volume to a non-muted value when needed to ensure
the conditions of the method are tested, not the side-effects
of only changing the ringer mode. This will also ensure the
assert is evaluated once the volume has changed.

Bug: 409665093
Flag: EXEMPT test improvement
Test: atest CtsMediaAudioTestCases:android.media.audio.cts.AudioFocusTest#testAudioFocusExclusive
Change-Id: I88cb027e141f7733fa6b73ab48bb7f837b617f9a